Move promise creator to utils.
