+++ src/js/app.ts
@@ -2,6 +2,7 @@ import { getSelector, getPuzzleType, getTileSelection, placeTile } from "./app-o
 import { DisplayManager } from "./puzzle-display";
 import { getTetrahedron, getTilePool } from "./puzzle-loader";
 import { Tetrahedron } from "./tetrahedron";
+import { createPromise } from "./utils";
 
 
 // Track tile placing event timer.
@@ -52,46 +53,6 @@ function testDisplay(): void {
             attachRotateEvents(puzzleDisplay, tetrahedron, displayManager);
         }
     }, 1000);
-}
-
-function createPromise(solver: (id: number, resolve: () => void, solving: number) => void, solving: number): { cancel: () => void; promise: Promise<unknown> } {
-    // Define completion flag and cancel trigger.
-    let finished = false;
-    let cancel = (): void => {
-        finished = true;
-    }
-    // Create the promise to solve the puzzle.
-    const promise = new Promise((resolve, reject) => {
-        const id = setInterval(() => {
-            solver(id, resolve, solving);
-        }, 1000);
-        // Triggering the cancel.
-        cancel = (): void => {
-            // If already completed no need to cancel.
-            if (finished) {
-                return;
-            }
-            // Process the cancel.
-            console.log("Cancelling...");
-            clearInterval(id);
-            reject();
-        }
-        // If cancelled before promise is started.
-        if (finished) {
-            cancel();
-        }
-    })
-        // Always set 'finished', so further cancelling has no effect.
-        // Note: Can't use finally here as we have targeted ES2016.
-        .then((resolvedValue) => {
-            finished = true;
-            return resolvedValue;
-        })
-        .catch((err) => {
-            finished = true;
-            return err;
-        });
-    return { promise, cancel }
 }
 
 function solvePuzzle(id: number, resolve: () => void, solving: number): void {
